方法整理
