--- src/com/ihgoo/allinone/util/EncryptUtils.java
@@ -1,36 +1,0 @@
-package com.ihgoo.allinone.util;
-
-import java.security.SecureRandom;
-
-import javax.crypto.Cipher;
-import javax.crypto.SecretKey;
-import javax.crypto.SecretKeyFactory;
-import javax.crypto.spec.DESKeySpec;
-
-/**
- * 加密工具类
- * @author Kelvin
- *
- */
-public class EncryptUtils {
-
-	/**
-	 * Des加密
-	 * @param private_key 私钥
-	 * @param plainText 明文密码
-	 * @return
-	 * @throws Exception
-	 */
-	private static byte[] desEncrypt(String private_key, byte[] plainText) throws Exception {
-		SecureRandom sr = new SecureRandom();
-		byte rawKeyData[] = private_key.getBytes();
-		DESKeySpec dks = new DESKeySpec(rawKeyData);
-		SecretKeyFactory keyFactory = SecretKeyFactory.getInstance("DES");
-		SecretKey key = keyFactory.generateSecret(dks);
-		Cipher cipher = Cipher.getInstance("DES");
-		cipher.init(Cipher.ENCRYPT_MODE, key, sr);
-		byte data[] = plainText;
-		byte encryptedData[] = cipher.doFinal(data);
-		return encryptedData;
-	}
-}
