+++ src/com/ihgoo/allinone/util/NetUtil.java
@@ -117,5 +117,6 @@ public class NetUtil {
         }
         return status;
     }
+    
 
 }
