+++ src/com/ihgoo/allinone/util/OtherUtils.java
@@ -210,128 +210,11 @@ public class OtherUtils {
 		return Thread.currentThread().getStackTrace()[4];
 	}
 
-	public static boolean checkFsWritable() {
-		// Create a temporary file to see whether a volume is really writeable.
-		// It's important not to put it in the root directory which may have a
-		// limit on the number of files.
-
-		// Logger.d(TAG, "checkFsWritable directoryName ==   "
-		// + PathCommonDefines.APP_FOLDER_ON_SD);
-		File directory = new File(Persistence.APP_FOLDER_ON_SD);
-		if (!directory.isDirectory()) {
-			if (!directory.mkdirs()) {
-				LogUtils.e("checkFsWritable directoryName 000  ");
-				return false;
-			}
-		}
-		File f = new File(Persistence.APP_FOLDER_ON_SD, ".probe");
-		try {
-			// Remove stale file if any
-			if (f.exists()) {
-				f.delete();
-			}
-			if (!f.createNewFile()) {
-				return false;
-			}
-			f.delete();
-			return true;
-		} catch (IOException ex) {
-			return false;
-		}
-	}
-
-	/**
-	 * 计算listview的高度,但子ListView每个Item必须是LinearLayout
-	 * 
-	 * @param listView
-	 */
-	public static void setListViewHeightBasedOnChildren(ListView listView) {
-		ListAdapter listAdapter = listView.getAdapter();
-		if (listAdapter == null) {
-			// pre-condition
-			return;
-		}
-
-		int totalHeight = 0;
-		for (int i = 0; i < listAdapter.getCount(); i++) {
-			View listItem = listAdapter.getView(i, null, listView);
-			listItem.measure(0, 0);
-			totalHeight += listItem.getMeasuredHeight() + 5;
-		}
-
-		ViewGroup.LayoutParams params = listView.getLayoutParams();
-		params.height = totalHeight
-				+ (listView.getDividerHeight() * (listAdapter.getCount() - 1));
-		listView.setLayoutParams(params);
-	}
-
-	public static boolean hasStorage() {
-		boolean hasStorage = false;
-		String str = Environment.getExternalStorageState();
-
-		if (str.equalsIgnoreCase(Environment.MEDIA_MOUNTED)) {
-			hasStorage = checkFsWritable();
-		}
-
-		return hasStorage;
-	}
-
-	public static int freeSpaceOnSd() {
-		int freeSize = 0;
-
-		if (hasStorage()) {
-			StatFs statFs = new StatFs(Persistence.APP_FOLDER_ON_SD);
-
-			try {
-				long nBlocSize = statFs.getBlockSize();
-				long nAvailaBlock = statFs.getAvailableBlocks();
-				freeSize = (int) (nBlocSize * nAvailaBlock / 1024 / 1024);
-
-			} catch (Exception e) {
-				e.printStackTrace();
-			}
-		}
-		return freeSize;
-	}
-
 	public static void updateFileTime(String dir, String fileName) {
 		File file = new File(dir, fileName);
 		long newModifiedTime = System.currentTimeMillis();
 		file.setLastModified(newModifiedTime);
 	}
-
-	public static boolean checkNet(Context context) {
-		ConnectivityManager manager = (ConnectivityManager) context
-				.getSystemService(Context.CONNECTIVITY_SERVICE);
-		NetworkInfo info = manager.getActiveNetworkInfo();
-		if (info != null) {
-			return true;
-		}
-		return false;
-	}
-
-	public static String getAPN(Context context) {
-		String apn = "";
-		ConnectivityManager manager = (ConnectivityManager) context
-				.getSystemService(Context.CONNECTIVITY_SERVICE);
-		NetworkInfo info = manager.getActiveNetworkInfo();
-
-		if (info != null) {
-			if (ConnectivityManager.TYPE_WIFI == info.getType()) {
-				apn = info.getTypeName();
-				if (apn == null) {
-					apn = "wifi";
-				}
-			} else {
-				apn = info.getExtraInfo().toLowerCase();
-				if (apn == null) {
-					apn = "mobile";
-				}
-			}
-		}
-		return apn;
-	}
-
 
 
 	private static SSLSocketFactory sslSocketFactory;
